Add Float type support and update benchmarks and tests

+++ benchmarks/rust/indicators/bench_bbands.rs
@@ -1,6 +1,7 @@
 use criterion::BenchmarkId;
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use techalysis::indicators::bbands::{bbands, BBandsMA};
+use techalysis::types::Float;
 
 fn bench_bbands(c: &mut criterion::Criterion) {
     let mut bench_group = c.benchmark_group("bbands");
@@ -9,7 +10,7 @@ fn bench_bbands(c: &mut criterion::Criterion) {
 
     for (len, period) in cases {
         let mut rng = StdRng::seed_from_u64(period as u64);
-        let data: Vec<f64> = (0..len).map(|_| rng.random_range(0.0..100.0)).collect();
+        let data: Vec<Float> = (0..len).map(|_| rng.random_range(0.0..100.0)).collect();
 
         bench_group.bench_with_input(
             BenchmarkId::new(format!("len={len}"), period),
