+++ crates/core/src/indicators/bbands.rs
@@ -3,43 +3,44 @@ use std::collections::VecDeque;
 use crate::errors::TechalysisError;
 use crate::indicators::ema::{ema_next_unchecked, period_to_alpha};
 use crate::indicators::sma::sma_next_unchecked;
+use crate::types::Float;
 
 #[derive(Debug)]
 pub struct BBandsResult {
-    pub upper_band: Vec<f64>,
-    pub middle_band: Vec<f64>,
-    pub lower_band: Vec<f64>,
+    pub upper_band: Vec<Float>,
+    pub middle_band: Vec<Float>,
+    pub lower_band: Vec<Float>,
     pub state: BBandsState,
 }
 
 #[derive(Debug, Clone)]
 pub struct BBandsState {
-    pub upper: f64,
-    pub middle: f64,
-    pub lower: f64,
-    pub sma: f64,
-    pub ma_sq: f64,
-    pub window: VecDeque<f64>,
+    pub upper: Float,
+    pub middle: Float,
+    pub lower: Float,
+    pub sma: Float,
+    pub ma_sq: Float,
+    pub window: VecDeque<Float>,
     pub period: usize,
-    pub std_up: f64,
-    pub std_down: f64,
+    pub std_up: Float,
+    pub std_down: Float,
     pub ma_type: BBandsMA,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq)]
 pub enum BBandsMA {
     SMA,
-    EMA(Option<f64>),
+    EMA(Option<Float>),
 }
 
-impl From<BBandsResult> for (Vec<f64>, Vec<f64>, Vec<f64>) {
+impl From<BBandsResult> for (Vec<Float>, Vec<Float>, Vec<Float>) {
     fn from(result: BBandsResult) -> Self {
         (result.upper_band, result.middle_band, result.lower_band)
     }
 }
 
 impl BBandsState {
-    pub fn next(&self, new_value: f64) -> Result<BBandsState, TechalysisError> {
+    pub fn next(&self, new_value: Float) -> Result<BBandsState, TechalysisError> {
         bbands_next(
             new_value,
             self.sma,
@@ -55,14 +56,14 @@ impl BBandsState {
 }
 
 pub fn bbands_next(
-    new_value: f64,
-    prev_sma: f64,
-    prev_ma: f64,
-    prev_ma_sq: f64,
-    window: &VecDeque<f64>,
+    new_value: Float,
+    prev_sma: Float,
+    prev_ma: Float,
+    prev_ma_sq: Float,
+    window: &VecDeque<Float>,
     period: usize,
-    std_up: f64,
-    std_down: f64,
+    std_up: Float,
+    std_down: Float,
     ma_type: BBandsMA,
 ) -> Result<BBandsState, TechalysisError> {
     if period <= 1 {
@@ -134,7 +135,7 @@ pub fn bbands_next(
             prev_ma_sq,
             std_up,
             std_down,
-            1.0 / period as f64,
+            1.0 / period as Float,
         ),
         BBandsMA::EMA(alpha) => {
             let alpha = if alpha.is_none() {
@@ -151,7 +152,7 @@ pub fn bbands_next(
                 alpha,
                 std_up,
                 std_down,
-                1.0 / period as f64,
+                1.0 / period as Float,
             )
         }
     };
@@ -171,10 +172,10 @@ pub fn bbands_next(
 }
 
 pub fn bbands(
-    data_array: &[f64],
+    data_array: &[Float],
     period: usize,
-    std_up: f64,
-    std_down: f64,
+    std_up: Float,
+    std_down: Float,
     ma_type: BBandsMA,
 ) -> Result<BBandsResult, TechalysisError> {
     let mut output_upper = vec![0.0; data_array.len()];
@@ -201,17 +202,17 @@ pub fn bbands(
 }
 
 pub fn bbands_into(
-    data: &[f64],
+    data: &[Float],
     period: usize,
-    std_up: f64,
-    std_down: f64,
+    std_up: Float,
+    std_down: Float,
     ma_type: BBandsMA,
-    output_upper: &mut [f64],
-    output_middle: &mut [f64],
-    output_lower: &mut [f64],
+    output_upper: &mut [Float],
+    output_middle: &mut [Float],
+    output_lower: &mut [Float],
 ) -> Result<BBandsState, TechalysisError> {
     let len = data.len();
-    let inv_period = 1.0 / (period as f64);
+    let inv_period = 1.0 / (period as Float);
     if period > len {
         return Err(TechalysisError::InsufficientData);
     }
@@ -322,14 +323,14 @@ pub fn bbands_into(
 
 #[inline(always)]
 pub fn bbands_sma_next_unchecked(
-    new_value: f64,
-    old_value: f64,
-    prev_ma: f64,
-    prev_ma_sq: f64,
-    std_up: f64,
-    std_down: f64,
-    inv_period: f64,
-) -> (f64, f64, f64, f64, f64) {
+    new_value: Float,
+    old_value: Float,
+    prev_ma: Float,
+    prev_ma_sq: Float,
+    std_up: Float,
+    std_down: Float,
+    inv_period: Float,
+) -> (Float, Float, Float, Float, Float) {
     let ma_sq = sma_next_unchecked(
         new_value * new_value,
         old_value * old_value,
@@ -343,45 +344,51 @@ pub fn bbands_sma_next_unchecked(
 
 #[inline(always)]
 pub fn bbands_ema_next_unchecked(
-    new_value: f64,
-    old_value: f64,
-    prev_sma: f64,
-    prev_ema: f64,
-    prev_sma_sq: f64,
-    alpha: f64,
-    std_up: f64,
-    std_down: f64,
-    inv_period: f64,
-) -> (f64, f64, f64, f64, f64) {
+    new_value: Float,
+    old_value: Float,
+    prev_sma: Float,
+    prev_ema: Float,
+    prev_sma_sq: Float,
+    alpha: Float,
+    std_up: Float,
+    std_down: Float,
+    inv_period: Float,
+) -> (Float, Float, Float, Float, Float) {
     let sma_sq = sma_next_unchecked(
         new_value * new_value,
         old_value * old_value,
         prev_sma_sq,
         inv_period,
     );
-    let sma: f64 = sma_next_unchecked(new_value, old_value, prev_sma, inv_period);
+    let sma: Float = sma_next_unchecked(new_value, old_value, prev_sma, inv_period);
     let middle = ema_next_unchecked(new_value, prev_ema, alpha);
     let (upper, lower) = bands(middle, sma, sma_sq, std_up, std_down);
     (upper, middle, lower, sma_sq, sma)
 }
 
 #[inline(always)]
-fn bands(middle: f64, mean: f64, mean_sq: f64, std_up: f64, std_down: f64) -> (f64, f64) {
+fn bands(
+    middle: Float,
+    mean: Float,
+    mean_sq: Float,
+    std_up: Float,
+    std_down: Float,
+) -> (Float, Float) {
     let std = (mean_sq - mean * mean).abs().sqrt();
     (middle + std_up * std, middle - std_down * std)
 }
 
 #[inline(always)]
 fn init_state_unchecked(
-    data: &[f64],
+    data: &[Float],
     period: usize,
-    inv_period: f64,
-    std_up: f64,
-    std_down: f64,
-    output_upper: &mut [f64],
-    output_middle: &mut [f64],
-    output_lower: &mut [f64],
-) -> Result<f64, TechalysisError> {
+    inv_period: Float,
+    std_up: Float,
+    std_down: Float,
+    output_upper: &mut [Float],
+    output_middle: &mut [Float],
+    output_lower: &mut [Float],
+) -> Result<Float, TechalysisError> {
     let (mut sum, mut sum_sq) = (0.0, 0.0);
     for idx in 0..period {
         let value = &data[idx];
@@ -394,9 +401,9 @@ fn init_state_unchecked(
             sum += value;
             sum_sq += value * value;
         }
-        output_upper[idx] = f64::NAN;
-        output_middle[idx] = f64::NAN;
-        output_lower[idx] = f64::NAN;
+        output_upper[idx] = Float::NAN;
+        output_middle[idx] = Float::NAN;
+        output_lower[idx] = Float::NAN;
     }
     output_middle[period - 1] = sum * inv_period;
     let ma_sq = sum_sq * inv_period;
